requirements.txt
# Data Collection
requests>=2.31.0
praw>=7.7.1
pandas>=2.2.0
beautifulsoup4>=4.12.2

# Sentiment Analysis
vaderSentiment>=3.3.2
openai>=1.3.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.1
python-dateutil>=2.8.2

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0